Move user notifications behavior to a concern
