+++ app/models/concerns/with_notifications.rb
@@ -1,0 +1,17 @@
+module WithNotifications
+  extend ActiveSupport::Concern
+
+  def unread_messages
+    messages.where read: false
+  end
+
+  def unread_notifications
+    # TODO: message and discussion should trigger a notification instead of being one
+    all = notifications.where(read: false) + unread_messages + unread_discussions
+    all.sort_by(&:created_at).reverse
+  end
+
+  def read_notification!(target)
+    notifications.find_by(target: target)&.mark_as_read!
+  end
+end
